pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "AppLib"
dynamic = ["version"]
description = "An Application Component Library (short AppLib) which makes it easy to create fancy GUIs"
readme = "README.md"
keywords = ["fluent-design", "PyQt6"]
requires-python = "==3.12.*"
license = { file = "LICENSE" }
authors = [{ name = "TheRealMorgenfrue" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.12",
]
dependencies = ["PyQt6-Fluent-Widgets==1.7.4", "qasync==0.28.0", "applib-core"]

[project.urls]
Homepage = "https://github.com/TheRealMorgenfrue/AppLib"
Documentation = "https://example.com"
Issues = "https://github.com/TheRealMorgenfrue/AppLib/issues"

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    { include-group = "docs" },
]
lint = ["ruff"]
test = ["pytest", "pytest-cov"]
docs = [
    "mkdocs-material",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-git-committers-plugin-2",
]

[tool.hatch.version]
path = "src/applib/version.py"

[tool.ruff]
extend-exclude = []

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
testpaths = ["tests/automatic"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.uv.workspace]
members = ["packages/applib-core"]

[tool.uv.sources]
applib-core = { workspace = true }
